Complete synchronous weather client

+++ src/main/java/sda/code/intermediate/part2/answers/json/gson/Coord.java
@@ -11,17 +11,17 @@ public class Coord {
 
     @SerializedName("lon")
     @Expose
-    private Integer lon;
+    private Double lon;
     @SerializedName("lat")
     @Expose
-    private Integer lat;
+    private Double lat;
 
     /**
      * 
      * @return
      *     The lon
      */
-    public Integer getLon() {
+    public Double getLon() {
         return lon;
     }
 
@@ -30,7 +30,7 @@ public class Coord {
      * @param lon
      *     The lon
      */
-    public void setLon(Integer lon) {
+    public void setLon(Double lon) {
         this.lon = lon;
     }
 
@@ -39,7 +39,7 @@ public class Coord {
      * @return
      *     The lat
      */
-    public Integer getLat() {
+    public Double getLat() {
         return lat;
     }
 
@@ -48,7 +48,7 @@ public class Coord {
      * @param lat
      *     The lat
      */
-    public void setLat(Integer lat) {
+    public void setLat(Double lat) {
         this.lat = lat;
     }
