+++ src/main/java/sda/code/intermediate/part2/answers/json/jackson/Coord.java
@@ -21,9 +21,9 @@ import org.apache.commons.lang3.builder.ToStringBuilder;
 public class Coord {
 
     @JsonProperty("lon")
-    private Integer lon;
+    private Double lon;
     @JsonProperty("lat")
-    private Integer lat;
+    private Double lat;
     @JsonIgnore
     private Map<String, Object> additionalProperties = new HashMap<String, Object>();
 
@@ -33,7 +33,7 @@ public class Coord {
      *     The lon
      */
     @JsonProperty("lon")
-    public Integer getLon() {
+    public Double getLon() {
         return lon;
     }
 
@@ -43,7 +43,7 @@ public class Coord {
      *     The lon
      */
     @JsonProperty("lon")
-    public void setLon(Integer lon) {
+    public void setLon(Double lon) {
         this.lon = lon;
     }
 
@@ -53,7 +53,7 @@ public class Coord {
      *     The lat
      */
     @JsonProperty("lat")
-    public Integer getLat() {
+    public Double getLat() {
         return lat;
     }
 
@@ -63,7 +63,7 @@ public class Coord {
      *     The lat
      */
     @JsonProperty("lat")
-    public void setLat(Integer lat) {
+    public void setLat(Double lat) {
         this.lat = lat;
     }
